Move RelatedRecordSelector component types to the ui-spec module
relatedTypeSchema, RELATED_RECORD_SELECTOR and the selector params
schema describe ui-spec component shapes, so they belong with the other
field-definition types. This also removes the uiSpecification to
databaseEngine import edge that sharing the child-field scan
introduced; the package export surface is unchanged via the barrels.
